+++ packages/stargate/scripts/define-proto.sh
@@ -33,6 +33,7 @@ protoc \
   "$COSMOS_PROTO_DIR/cosmos/staking/v1beta1/tx.proto" \
   "$COSMOS_PROTO_DIR/cosmos/tx/signing/v1beta1/signing.proto" \
   "$COSMOS_PROTO_DIR/cosmos/tx/v1beta1/tx.proto" \
+  "$COSMOS_PROTO_DIR/cosmos/vesting/v1beta1/vesting.proto" \
   "$COSMOS_PROTO_DIR/ibc/core/channel/v1/channel.proto" \
   "$COSMOS_PROTO_DIR/ibc/core/channel/v1/query.proto" \
   "$COSMOS_PROTO_DIR/ibc/core/client/v1/client.proto" \