+++ packages/stargate/src/stargateclient.ts
@@ -5,21 +5,18 @@ import {
   isSearchByHeightQuery,
   isSearchBySentFromOrToQuery,
   isSearchByTagsQuery,
-  PubKey,
   SearchTxFilter,
   SearchTxQuery,
 } from "@cosmjs/launchpad";
-import { Uint53, Uint64 } from "@cosmjs/math";
-import { decodePubkey } from "@cosmjs/proto-signing";
+import { Uint53 } from "@cosmjs/math";
 import {
   broadcastTxCommitSuccess,
   Tendermint34Client,
   toRfc3339WithNanoseconds,
 } from "@cosmjs/tendermint-rpc";
-import { assert, assertDefinedAndNotNull } from "@cosmjs/utils";
-import Long from "long";
+import { assertDefinedAndNotNull } from "@cosmjs/utils";
 
-import { BaseAccount } from "./codec/cosmos/auth/v1beta1/auth";
+import { Account, accountFromAny } from "./accounts";
 import { MsgData, TxMsgData } from "./codec/cosmos/base/abci/v1beta1/abci";
 import { Coin } from "./codec/cosmos/base/v1beta1/coin";
 import { AuthExtension, BankExtension, QueryClient, setupAuthExtension, setupBankExtension } from "./queries";
@@ -33,14 +30,6 @@ export interface IndexedTx {
   readonly code: number;
   readonly rawLog: string;
   readonly tx: Uint8Array;
-}
-
-export interface Account {
-  /** Bech32 account address */
-  readonly address: string;
-  readonly pubkey: PubKey | null;
-  readonly accountNumber: number;
-  readonly sequence: number;
 }
 
 export interface SequenceResponse {
@@ -84,23 +73,6 @@ export function assertIsBroadcastTxSuccess(
       `Error when broadcasting tx ${result.transactionHash} at height ${result.height}. Code: ${result.code}; Raw log: ${result.rawLog}`,
     );
   }
-}
-
-function uint64FromProto(input: number | Long | null | undefined): Uint64 {
-  if (!input) return Uint64.fromNumber(0);
-  return Uint64.fromString(input.toString());
-}
-
-export function accountFromProto(input: BaseAccount): Account {
-  const { address, pubKey, accountNumber, sequence } = input;
-  const pubkey = decodePubkey(pubKey);
-  assert(address);
-  return {
-    address: address,
-    pubkey: pubkey,
-    accountNumber: uint64FromProto(accountNumber).toNumber(),
-    sequence: uint64FromProto(sequence).toNumber(),
-  };
 }
 
 export function coinFromProto(input: Coin): Coin {
@@ -151,14 +123,14 @@ export class StargateClient {
   // this is nice to display data to the user, but is slower
   public async getAccount(searchAddress: string): Promise<Account | null> {
     const account = await this.queryClient.auth.account(searchAddress);
-    return account ? accountFromProto(account) : null;
+    return account ? accountFromAny(account) : null;
   }
 
   // if we just need to get the sequence for signing a transaction, let's make this faster
   // (no need to wait a block before submitting)
   public async getAccountUnverified(searchAddress: string): Promise<Account | null> {
     const account = await this.queryClient.auth.unverified.account(searchAddress);
-    return account ? accountFromProto(account) : null;
+    return account ? accountFromAny(account) : null;
   }
 
   public async getSequence(address: string): Promise<SequenceResponse | null> {